Adding key press callbacks instead of setting it.
Calling set(figHandle, 'WindowKeyPressFcn', @callback) as it was done previously would completely kill any previously-defined callbacks on the figure. Adding callbacks via iptaddcallback is cleaner. Unfortunately, this also means we keep adding callbacks if imshowHDR is called repetitively on the same axes! Care is taken to ensure that callbacks for the same axis handles are not added twice.